+++ Classes/Reference+Resolving.swift
@@ -22,68 +22,68 @@ extension Reference
 		- returns: An instance of the desired type, or nil if it hasn't been resolved yet OR is of a different type
 	 */
 	public func resolved<T: Resource>(type: T.Type) -> T? {
-		let refid = processedReferenceIdentifier()
-		if nil == refid {
-			print("Reference.resolved(): This reference does not have a reference-id, cannot resolve")
+		guard let refid = processedReferenceIdentifier() else {
+			fhir_warn("reference \(self) does not have a reference-id, cannot resolve")
 			return nil
 		}
 		
-		if let resolved = _owner?.resolvedReference(refid!) {
+		if let resolved = _owner?.resolvedReference(refid) {
 			if let res = resolved as? T {
 				return res
 			}
-			print("Reference.resolved(): Reference \(refid) was dereferenced to \(resolved), which is not of the expected type \(T.self)")
+			fhir_warn("reference “\(refid)” was dereferenced to «\(resolved)», which is not of the expected type “\(T.self)”")
 		}
 		
 		// not yet resolved, let's look at contained resources
-		if let contained = _owner?.containedReference(refid!) {
-			let t = T.self									// getting crashes when using T(...) directly as of 6.1 GM 2
-			let instance = t.init(json: contained.json)
-			contained.owner.didResolveReference(refid!, resolved: instance)
-			return instance
+		if let contained = _owner?.containedReference(refid) {
+			return contained.resolvedInstanceWithReferenceId(refid, ofType: type)
 		}
-		
 		return nil
 	}
 	
 	/**
-		Checks if a reference can be resolved immediately by calling `resolved()` first, if not proceeds to request the
-		referenced resource from the respective location.
+	Checks if a reference can be resolved immediately by calling `resolved()` first, if not proceeds to request the referenced resource from
+	the respective location.
 	
-		- parameter type: The type of the resource to expect
-		- parameter callback: The callback to call upon success or failure, with the resolved resource or nil
-	 */
+	- parameter type: The type of the resource to expect
+	- parameter callback: The callback to call upon success or failure, with the resolved resource or nil
+	*/
 	public func resolve<T: Resource>(type: T.Type, callback: (T? -> Void)) {
 		if let resolved = resolved(T) {
 			callback(resolved)
 		}
 		else if let ref = reference {
-			if let server = owningResource()?._server {
-				
-				// TODO: absolute URL
-				if let _ = ref.rangeOfString("://") {
-					print("TODO: Resolve reference with absolute URL \(ref)")
-					callback(nil)
-				}
-					
-				// relative URL
-				else {
-					T.readFrom(ref, server: server) { resource, error in
-						if let res = resource {
-							self._owner?.didResolveReference(ref, resolved: res)
-							callback(res as? T)		// `readFrom()` will always instantiate its own type, so this should never fail
+			var server: FHIRServer? = nil
+			var path = ref
+			
+			// absolute URL
+			if let _ = ref.rangeOfString("://") {
+				path = (ref as NSString).lastPathComponent
+				let base = NSURL(string: (ref as NSString).stringByDeletingLastPathComponent)!
+				server = FHIRBaseServer(baseURL: base, auth: nil)		// TODO: what if it's protected?
+			}
+			
+			// relative URL
+			else if let srv = owningResource()?._server {
+				server = srv
+			}
+			
+			if let server = server {
+				T.readFrom(path, server: server) { resource, error in
+					if let res = resource {
+						self._owner?.didResolveReference(ref, resolved: res)
+						callback(res as? T)		// `readFrom()` will always instantiate its own type, so this should never fail
+					}
+					else {
+						if let err = error {
+							fhir_warn("error resolving reference «\(ref)»: \(err)")
 						}
-						else {
-							if let err = error {
-								print("Reference.resolve(): Error resolving reference \(ref): \(err)")
-							}
-							callback(nil)
-						}
+						callback(nil)
 					}
 				}
 			}
 			else {
-				print("Reference.resolve(): The reference \(self) does not have a server instance, cannot resolve")
+				fhir_warn("reference \(self) does not have a server instance, cannot resolve «\(ref)»")
 				callback(nil)
 			}
 		}