+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Changelog
 ## 2.2
 
 - Update to FHIR `1.0.2.7202` (DSTU 2 with technical errata, compatible to `1.0.1`)
+- Add a very simple base implementation of `FHIRServer` called `FHIROpenServer`and a typealias `FHIRBaseServer`
 - New error handling using `FHIRError`
+- Implement absolute reference resolver (will not work if the other server is protected)
 - Fixes to `ElementDefinition`
 
 ## 2.1